requirements.txt
amqp==2.4.0
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
confusable-homoglyphs==3.2.0
coverage==4.5.1
dj-database-url==0.5.0
Django==2.1.3
django-admin-row-actions==0.0.4
django-appconf==1.0.2
django-cleanup==3.0.1
django-extensions==2.1.5
django-filter==2.0.0
django-google-maps==0.10.1
django-imagekit==4.0.2
django-jet==1.0.8
django-phonenumber-field==2.1.0
django-registration==3.0
django-reversion==3.0.2
flower==0.9.2
gunicorn==19.9.0
kombu==4.2.2.post1
more-itertools==4.3.0
phonenumberslite==8.9.16
pilkit==2.0
Pillow==5.1.0
pluggy==0.8.0
pocketsphinx==0.1.15
pprintpp==0.4.0
psycopg2==2.7.5
py==1.7.0
PyAudio==0.2.11
pytest==3.10.0
pytest-cov==2.6.0
pytest-mock==1.10.0
python-dateutil==2.7.5
python-http-client==3.1.0
pytz==2018.7
rc-django-celery-results==1.0.1
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.11.0
sox==1.3.3
vine==1.2.0
whitenoise==4.1
